ocr/PaddleOCR/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.0
astor==0.8.1
autocorrect==2.6.1
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.2
decorator==5.2.1
easygui==0.98.3
fire==0.7.1
fonttools==4.61.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
matplotlib==3.10.7
networkx==3.6
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.9.6.50
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.10.0
paddlepaddle-gpu==3.0.0
pillow==12.0.0
protobuf==6.33.1
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
pyspellchecker==0.8.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.3.1
shapely==2.1.2
simsimd==6.5.3
six==1.17.0
soupsieve==2.8
stringzilla==4.4.0
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
validate_docbr==1.11.1
wheel==0.45.1